Added current logged in user message in UI
This commit includes a new function 'returnCurrentUser' in the UI class, which can be reused anywhere to print out the user who is currently logged in. It has to be a function, as this allows it to use a try and catch block to prevent crashes if the user has not been logged in yet if the function somehow gets misplaced. The program won't compile either if one tries to use values that need to be user generated in a variable.